Made the port ID of the server be taken as a 32-bit int, and checked agains 16-bit max.

+++ project/server/src/main.cpp
@@ -12,7 +12,7 @@ void usage() {
 }
 
 uint16_t parseArgs(int argc, char** argv) {
-    uint16_t portId;
+    uint32_t portId;
     if (argc == 1) {
         portId = 80;
     }
